Member

+++ backend/src/main/java/com/ourdressingtable/member/service/impl/MemberServiceImpl.java
@@ -5,6 +5,7 @@ import com.ourdressingtable.exception.OurDressingTableException;
 import com.ourdressingtable.member.domain.ColorType;
 import com.ourdressingtable.member.domain.Member;
 import com.ourdressingtable.member.domain.SkinType;
+import com.ourdressingtable.member.domain.Status;
 import com.ourdressingtable.member.domain.WithdrawalMember;
 import com.ourdressingtable.member.dto.CreateMemberRequest;
 import com.ourdressingtable.member.dto.OtherMemberResponse;
@@ -13,6 +14,8 @@ import com.ourdressingtable.member.dto.WithdrawalMemberRequest;
 import com.ourdressingtable.member.repository.MemberRepository;
 import com.ourdressingtable.member.repository.WithdrawalMemberRepository;
 import com.ourdressingtable.member.service.MemberService;
+import com.ourdressingtable.util.HashUtil;
+import com.ourdressingtable.util.MaskingUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -53,19 +56,29 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     @Transactional
-    public void deleteMember(Long id) {
+    public void deleteMember(Long id, WithdrawalMemberRequest withdrawalMemberRequest) {
         Member member = memberRepository.findById(id).orElseThrow(() -> new OurDressingTableException(ErrorCode.MEMBER_NOT_FOUND));
-        memberRepository.delete(member);
+        Long newId = createWithdrawalMember(withdrawalMemberRequest, member);
+        if(newId != null) {
+            memberRepository.delete(member);
+        }
     }
 
     @Override
     @Transactional
-    public Long createWithdrawMember(Long id, WithdrawalMemberRequest withdrawalMemberRequest) {
-        Member member = memberRepository.findById(id).orElseThrow(() -> new OurDressingTableException(ErrorCode.MEMBER_NOT_FOUND));
-        member.changeMemberStatus();
-
-        WithdrawalMember withdrawalMember = withdrawalMemberRepository.save(withdrawalMemberRequest.toEntity());
+    public Long createWithdrawalMember(WithdrawalMemberRequest withdrawalMemberRequest, Member member) {
+        String maskedEmail = MaskingUtil.maskedEmail(member.getEmail());
+        String maskedPhone = MaskingUtil.maskedPhone(member.getPhoneNumber());
+        String hashedEmail = HashUtil.hash(member.getEmail());
+        String hashedPhone = HashUtil.hash(member.getPhoneNumber());
+        boolean isBlock = false;
+        if(member.getStatus() == Status.BLOCK) {
+            isBlock = true;
+        }
+        WithdrawalMember withdrawalMember = withdrawalMemberRepository.save(withdrawalMemberRequest.toEntity(hashedEmail,hashedPhone,maskedEmail,maskedPhone, isBlock));
         return withdrawalMember.getId();
-
     }
+
+
+
 }
